Migrate upgrade-cluster
