--- web/html/src/manager/clusters/upgrade-cluster/upgrade-cluster.js
@@ -1,81 +1,0 @@
-// @flow
-import { hot } from 'react-hot-loader/root';
-import * as React from 'react';
-import { useState } from 'react';
-import { TopPanel } from "components/panels/TopPanel";
-import { HashRouter, Route, Switch } from 'components/utils/HashRouter';
-import withPageWrapper from 'components/general/with-page-wrapper';
-import useClustersApi, { withErrorMessages } from '../shared/api/use-clusters-api';
-import ScheduleClusterAction from '../shared/ui/schedule-cluster-action';
-import FormulaConfig from '../shared/ui/formula-config';
-import UpgradeClusterPlan from './upgrade-cluster-plan';
-
-import type { FormulaValuesType } from '../shared/api/use-clusters-api';
-import type { ClusterType } from '../shared/api/use-clusters-api';
-
-type Props = {
-    cluster: ClusterType,
-    flashMessage: string,
-    showUpgradePlan: boolean
-};
-
-const UpgradeCluster = (props: Props) => {
-    const [upgradeConfig, setUpgradeConfig] = useState<?FormulaValuesType>(null);
-
-    const { scheduleUpgradeCluster } = useClustersApi();
-
-    const scheduleUpgrade = (earliest: Date, actionChain: ?string): Promise<any> => {
-        return scheduleUpgradeCluster(props.cluster.id, earliest, actionChain)
-                .then(() => {
-                    window.location = `/rhn/manager/cluster/${props.cluster.id}`;
-                });
-    }
-
-    return (<TopPanel title={t('Upgrade ') + props.cluster.name}
-        icon="spacewalk-icon-clusters"
-        helpUrl="reference/clusters/clusters-menu.html">
-        <HashRouter initialPath={props.showUpgradePlan ? "upgrade-plan" : "upgrade-config"}>
-            <Switch>
-                <Route path="upgrade-plan">
-                    {({ goTo, back }) =>
-                        <UpgradeClusterPlan
-                            cluster={props.cluster}
-                            onNext={(formulaValues) => { goTo("upgrade-config"); }}
-                        />}
-                </Route>            
-                <Route path="upgrade-config">
-                    {({ goTo, back }) =>
-                        <FormulaConfig title={t("Configuration override")}
-                            values={upgradeConfig}
-                            formula="upgrade_cluster"
-                            context={{ "cluster": props.cluster.id }}
-                            provider={props.cluster.provider.label}
-                            onPrev={props.showUpgradePlan ? back: null}
-                            onNext={(formulaValues) => { setUpgradeConfig(formulaValues); goTo("schedule"); }}
-                        />}
-                </Route>
-                <Route path="schedule">
-                    {({ goTo, back }) => {
-                        return upgradeConfig ? <ScheduleClusterAction
-                            title={t("Schedule upgrade cluster")}
-                            panel={
-                                <div className="form-horizontal">
-                                    <div className="form-group">
-                                        <label className="col-md-3 control-label">{t("Cluster:")}</label>
-                                        <div className="col-md-9">{props.cluster.name}</div>
-                                    </div>
-                                </div>
-                            }
-                            schedule={scheduleUpgrade}
-                            scheduleButtonLabel={t("Upgrade")}
-                            onPrev={back}
-                            actionType="cluster.upgrade_cluster"
-                        /> : goTo()
-                    }}
-                </Route>
-            </Switch>
-        </HashRouter>
-    </TopPanel>);
-}
-
-export default hot(withPageWrapper<Props>(withErrorMessages(UpgradeCluster)));
